structuring the database layer
